expose internal oas in non-prod envs, reorganize trpc
