+++ packages/commonwealth/package.json
@@ -147,7 +147,6 @@
     "@trpc/client": "^10.45.1",
     "@trpc/react-query": "^10.45.1",
     "@types/react-helmet-async": "^1.0.3",
-    "@types/swagger-ui-express": "^4.1.6",
     "@viem/anvil": "^0.0.10",
     "@walletconnect/ethereum-provider": "^2.10.1",
     "@walletconnect/modal": "^2.4.6",
@@ -251,7 +250,6 @@
     "stream-browserify": "^3.0.0",
     "stream-http": "^3.2.0",
     "superagent": "^5.2.2",
-    "swagger-ui-express": "^5.0.0",
     "tsconfig-paths": "^4.2.0",
     "tsx": "^4.7.2",
     "tweetnacl": "^1.0.3",
