+++ packages/commonwealth/server/api/index.ts
@@ -1,2 +1,3 @@
 export * as external from './external-router';
+export * as internal_oas from './internal-oas';
 export * as internal from './internal-router';
